some api changes; enabled ogg vorbis plugin
git-svn-id: svn://svn.code.sf.net/p/qmmp-dev/code/trunk@565 90c681e8-e032-0410-971d-27865f9a5e38

+++ qmmp/src/plugins/Input/mad/decoder_mad.h
@@ -28,7 +28,7 @@ public:
     // standard decoder API
     bool initialize();
     qint64 lengthInSeconds();
-    void seek(double);
+    void seek(qint64);
     void stop();
 
 private:
@@ -44,7 +44,7 @@ private:
     bool findHeader();
     bool findXingHeader(struct mad_bitptr, unsigned int);
     bool inited, user_stop, done, m_finish, derror, eof, useeq;
-    double totalTime, seekTime;
+    qint64 totalTime, seekTime;
     int channels;
     unsigned long bitrate;
     long freq, len;
